requirements.txt
accelerate==0.25.0
aiohttp==3.9.0
aiosignal==1.3.1
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
backend==0.2.4.1
beautifulsoup4==4.12.2
black==23.12.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ctransformers==0.2.27
cv==1.0.0
dataclasses-json==0.6.2
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
easygui==0.98.3
et-xmlfile==1.1.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.20.2
idna==3.4
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.340
langchain-community==0.0.10
langchain-core==0.1.8
langsmith==0.0.66
loguru==0.7.2
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
openai==1.3.5
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pdf2image==1.17.0
peppercorn==0.6
Pillow==10.1.0
pinecone-client==2.2.4
pipenv==2023.11.15
piptoken==0.0.1
platformdirs==4.0.0
protobuf==4.25.1
psutil==5.9.7
py-cpuinfo==9.0.0
pyarrow==14.0.2
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pypdf==3.17.4
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
replicate==0.22.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
safetensors==0.4.1
serpapi==0.1.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
virtualenv==20.24.7
yarl==1.9.3
zipp==3.17.0
